sync: extract SyncDatabase so the SQLite handle is file-private AND files split
Resolves the tension between the two prior review preferences: a +SQLite
extension widened the raw handle to module-internal (actor-isolation
weakening), while keeping one 619-line file tripped the file-length guard.
Extract a SyncDatabase type that owns the raw sqlite3 OpaquePointer as a
PRIVATE member and exposes only the binder/exec/transaction/prepare helpers;
CmuxSyncStore holds one as a private let. The handle is now never
module-visible (isolation invariant holds by construction), and the package
splits into focused files (store 492, SyncDatabase 120, row-decode 32) so no
budget exception is needed. Pure code-motion + indirection; behavior
unchanged, all 49 package tests green.
